Import ansible version 2.1.0.0 as sysutils/ansible2.
This is based on the existing sysutils/ansible package, but cleaned up and
tracking the newer 2.x releases.  sysutils/ansible is retained to support
installations which still require 1.x support.  From the DESCR:

Ansible is a radically simple model-driven configuration management,
multi-node deployment, and remote task execution system. Ansible works
over SSH and does not require any software or daemons to be installed
on remote nodes. Extension modules can be written in any language and
are transferred to managed machines automatically.