Changes 5.5.51: Bugs Fixed
Replication: When using statement-based or mixed binary logging format with --read-only=ON, it was not possible to modify temporary tables.

MySQL Server upgrades performed using RPM packages failed when upgrading from MySQL 5.1 Community to MySQL 5.5 Community or MySQL 5.1 Commercial to MySQL 5.5 Commercial.

A buffer overflow in the regex library was fixed.

Certain arguments to NAME_CONST() could cause a server exit.

Installing MySQL from a yum or zypper repository resulted in /var/log/mysqld.log being created with incorrect user and group permissions.

If a stored function updated a view for which the view table had a trigger defined that updated another table, it could fail and report an error that an existing table did not exist.

If an INSTALL PLUGIN statement contained invalid UTF-8 characters in the shared library name, it caused the server to hang (or to raise an assertion in debug builds).

For multibyte character sets, LOAD DATA could fail to allocate space correctly and ignore input rows as a result.