Extend check for gcc to work with distcc/ccache
Instead of checking that PKGSRC_COMPILER is equal to gcc, see if gcc
is contained in it.  Thus, if PKGSRC_COMPILER is "ccache gcc" or
"ccache distcc gcc", the hack will still be applied.  (As a side
effect, I can confirm that the hack is still needed on recent netbsd-7
i386.)